Move the table filters storing into a hook that can more easily handle the url parameter setting
